% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_summary}
\alias{calculate_summary}
\title{Calculate summary statistics (probabilities)}
\usage{
calculate_summary(samples)
}
\arguments{
\item{samples}{A binned statistics data frame.}
}
\value{
A data frame with probabilities
}
\description{
Aggregates bin statistics to region high probability risks. Requires columns
\emph{PPDX}, \emph{PPDX_norm} and \emph{over_mean}.
}
\details{
\enumerate{
\item Sums sample \emph{PPDX}, \emph{PPDX_norm}, \emph{over_mean}.
\item Calculates reference mean of \emph{PPDX} and \emph{PPDX_norm}.
\item For each subregion (focus) calculates Mahalanobis distance from the
reference group (\emph{PPDX}, \emph{PPDX_norm}) and transforms it with chi-square(df = 2)
to -log10 probability.
}
}
