% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_reference}
\alias{check_reference}
\title{Sanity check for the reference data frame (after initial read in).}
\usage{
check_reference(reference, use_pca, nComp)
}
\arguments{
\item{reference}{A data frame holding reference set}

\item{use_pca}{Is PCA-normalization switched on?}

\item{nComp}{How many PCA components are used?}
}
\description{
Checks if:
\enumerate{
\item required column names are present.
\item reference sample size is not 0.
\item number of PCA components suit with reference sample size.
}
}
