% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_uniq}
\alias{check_uniq}
\title{Check that analyzable sample name is not present in reference set.}
\usage{
check_uniq(binned_reads, reference)
}
\arguments{
\item{binned_reads}{binned BAM}

\item{reference}{A data frame holding reference set}
}
\description{
This check is important as otherwise using group_by in the metrics
calculations will lead to incorrect calculations.
}
