% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{divide_bins}
\alias{divide_bins}
\title{Create based on the location file.}
\usage{
divide_bins(locations_file)
}
\arguments{
\item{locations_file}{A tab separated text file describing how to create bins with columns: \emph{chr}, \emph{start}, \emph{end}, \emph{focus}, \emph{length}.}
}
\value{
A data frame containing bins.
}
\description{
Create based on the location file.
}
