% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{gc_correct}
\alias{gc_correct}
\title{Apply GC\% correction to samples.}
\usage{
gc_correct(samples)
}
\arguments{
\item{samples}{A data frame containing the required columns.}
}
\value{
A data frame with an added column \emph{gc_corrected}.
}
\description{
This function applies GC\% correction to the read counts in the given samples data frame.
It assumes the presence of columns: \emph{chr}, \emph{start}, \emph{end}, \emph{focus}, \emph{reads}, and \emph{sample}.
}
