% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindel.R
\name{infer_normality}
\alias{infer_normality}
\title{Infer sample high-risk probability based on a reference set.}
\usage{
infer_normality(
  bam_file_path,
  ref_file_path,
  cumulative_variance = 95,
  cumulative_variance_per_region = NULL,
  output_file_path = NULL,
  bin_file_path = NULL
)
}
\arguments{
\item{bam_file_path}{character. The path to the BAM file to be processed.}

\item{ref_file_path}{character. The path to the reference file in TSV format.}

\item{cumulative_variance}{numeric, optional. The cumulative variance threshold (\%) for general PCA correction over all the genome. Default is 95 (\%).}

\item{cumulative_variance_per_region}{numeric vector, optional. A vector of cumulative variance thresholds (\%) for PCA correction to apply per region. Default is NULL.}

\item{output_file_path}{character, optional. The path to the output file where results will be saved. If not provided, results will be returned as a data frame. Default is NULL.}

\item{bin_file_path}{character, optional. The path to the file where bin data will be saved. Default is NULL.}
}
\value{
If \code{output_file_path} is not provided, returns a data frame with the processed data. Otherwise, results are saved to the specified output file.
}
\description{
Infer sample high-risk probability based on a reference set.
}
\examples{
infer_sample_high_risk_probability(
  bam_file_path = "sample.bam",
  ref_file_path = "reference.tsv",
  cumulative_variance = 95,
  cumulative_variance_per_region = c(0, 10),
  output_file_path = "output.tsv"
)
}
