% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalize_reads}
\alias{normalize_reads}
\title{Normalize bin read count by sample total read count and bin length.}
\usage{
normalize_reads(samples)
}
\arguments{
\item{samples}{\code{\link{gc_correct}} output.}
}
\value{
A normalized data frame.
}
\description{
Normalize bin read count by sample total read count and bin length.
}
