% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{save_bin_plot}
\alias{save_bin_plot}
\title{Creates and saves binned normalized Z-Score figure per region.}
\usage{
save_bin_plot(samples, sample_name, file_name_prefix)
}
\arguments{
\item{sample_name}{sample name of interest}

\item{file_name_prefix}{name of the file to write}

\item{data}{A binned Z-scored data frame with columns \emph{focus}, \emph{start},
\emph{PPDX_norm}, \emph{sample} and \emph{reference}.}
}
\description{
Creates and saves binned normalized Z-Score figure per region.
}
