% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{write_reference}
\alias{write_reference}
\title{Create a BinDel reference}
\usage{
write_reference(
  bam_locations,
  coordinates_file,
  output_name,
  col_names = TRUE,
  anonymise = TRUE,
  prefix = "S"
)
}
\arguments{
\item{bam_locations}{A vector of file paths to BAM files that will be used to generate the reference.}

\item{coordinates_file}{A file path to the coordinates file. The file should contain the following columns: \emph{chr}, \emph{start}, \emph{end}, \emph{focus}, \emph{length}.}

\item{output_name}{The name of the output reference file. If the file extension \code{.gz} is appended to the name, the resulting file will be compressed.}

\item{col_names}{A logical value that specifies whether to include column names in the output file.}

\item{anonymise}{A logical value that specifies whether to include the original sample name in the reference file. If \code{TRUE}, the sample names are replaced with a \code{prefix + integer}.}

\item{prefix}{A character string that is used as a prefix when anonymising the sample names.}
}
\description{
This function creates a BinDel reference from a set of input BAM files and a coordinates file.
}
\details{
Note: The function assumes that the input BAM files and the coordinates file have already been prepared and are available for use in the function.
}
\examples{
write_reference(c("sample1.bam", "sample2.bam", "sample3.bam"), "locations.info.tsv", "reference.gz")
write_reference(c("sample.bam"), "locations.info.tsv", "reference.tsv", col_names = FALSE)
}
