% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_bin_stat}
\alias{calculate_bin_stat}
\title{Calculate bin statistics (microduplication and microdeletion risk bin stat).}
\usage{
calculate_bin_stat(samples)
}
\arguments{
\item{samples}{A normalized data frame containing the required columns.}
}
\value{
A data frame with additional columns \emph{dup_score} and \emph{md_score}.
}
\description{
Calculate statistics for each genomic bin compared to a reference group.
This function should be called after normalization.
}
\details{
Required columns in the input data frame are: \emph{chr}, \emph{start},
\emph{reference}, \emph{gc_corrected}.
}
