% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_summary}
\alias{calculate_summary}
\title{Calculate summary statistics including Mahalanobis distances.}
\usage{
calculate_summary(samples)
}
\arguments{
\item{samples}{A binned statistics data frame.}
}
\value{
A data frame with probabilities
}
\description{
Calculate summary statistics including Mahalanobis distances.
}
