% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_reference}
\alias{check_reference}
\title{Check reference data frame for consistency}
\usage{
check_reference(reference, use_pca, num_comp)
}
\arguments{
\item{reference}{A data frame holding reference set}

\item{use_pca}{Should PCA normalization be used?}

\item{num_comp}{The number of PCA components to use?}
}
\description{
This function checks the reference data frame for consistency, ensuring that it has the
expected columns, the sample size is not zero, and the number of PCA components is appropriate.
}
