% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{gc_correct}
\alias{gc_correct}
\title{Correct GC\% bias in bins}
\usage{
gc_correct(samples)
}
\arguments{
\item{samples}{A data frame to GC\% correct.}
}
\value{
A GC\% corrected data frame.
}
\description{
Sample wise, based on PMID: 28500333 and PMID: 20454671. Expects columns
\emph{chr}, \emph{start}, \emph{end}, \emph{focus}, \emph{reads},
\emph{samples}. Creates column \emph{gc_correct}.
}
