% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindel.R
\name{infer_normality}
\alias{infer_normality}
\title{Infer sample high-risk probability.}
\usage{
infer_normality(
  bam_location,
  reference_location,
  use_pca = TRUE,
  nComp = 80,
  bin_plot = TRUE,
  save_bins = FALSE
)
}
\arguments{
\item{bam_location}{Path to the input file.}

\item{reference_location}{Path to the reference file.}

\item{use_pca}{Use PCA based normalization? Increases sensitivity by reducing variation between samples.}

\item{nComp}{How many components to use in PCA-based normalization? Cannot be higher than the number of samples in the reference set. It affects the minimum read count and the fetal fraction that the tool can reliably operate on.}

\item{bin_plot}{Create and save detailed bin plots?}

\item{save_bins}{Save bins?}
}
\description{
This function does the following:
\enumerate{
\item Apply bin-based GC\% correct.
\item Normalizing by read count.
\item Normalize by bin length.
\item Apply PCA-based normalization
\item Calculate reference group statistics per focus region.
\item Calculate sample Mahalanobis distance from the reference group.
\item Calculate -log10 chi-squared distribution probabilities.
}
}
\details{
Outputs results to different files.
}
\examples{

bam <- "sample.bam"
reference <- "reference.tsv"
infer_normality(bam, reference)
head("sample.bam.tsv")
}
