% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{pca_correct}
\alias{pca_correct}
\title{Perform PCA-based normalization on a normalized data frame.}
\usage{
pca_correct(samples, cumulative_variance)
}
\arguments{
\item{samples}{A data frame resulting from the \code{\link{normalize_reads}} function.}

\item{cumulative_variance}{The cumulative variance threshold (\%) to determine the number of PCA components.}
}
\value{
A normalized data frame.
}
\description{
This function applies PCA-based normalization to a data frame that has undergone
previous normalization steps such as read count normalization and GC-correction.
}
\references{
This function is based on a \href{https://stats.stackexchange.com/questions/229092/how-to-reverse-pca-and-reconstruct-original-variables-from-several-principal-com}{StackExchange answer}.
}
